+++ CHANGELOG.md
@@ -9,12 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - Add support for Amazon Linux and restore Ubuntu 16.04 support (glibc 2.23)
+ 
 ### Changed
  - Fix `gdal.RasterBand.flushAsync()`
  - Implement the previously non-existing `strict` argument of `gdal.Driver.createCopy{Async}()` and move it after `options`
  - Implement progress callbacks in `gdal.Driver.createCopy{Async}()`
  - `gdal.Driver.createCopy{Async}()` now always throws an `Error` / rejects the Promise if the operation failed instead of returning a null object in some cases
  - Fix #14, correctly handle exceptions in progress callbacks and return them to the calling user code
+ - Add a previously missing TypeScript signature for `gdal.CoordinateTransformation.transformPoint(point: xyz)`
 
 ## [3.3.3] 2021-09-07